fix(web): move html/body shell into [locale]/layout to enable static rendering
Root layout's await getLocale() runs above the [locale] segment, before
setRequestLocale fires, so it falls back to reading request data and
opts the entire app out of static prerendering. Symptoms before the fix:
0 prerendered routes in prerender-manifest, 0 .html/.rsc files for any
[locale]/... page; everything served as dynamic.

Move <html>/<body>, fonts, providers, JSON-LD scripts, generateMetadata,
and viewport into [locale]/layout.tsx, where setRequestLocale(locale)
already runs first. Root app/layout.tsx becomes a pass-through that just
imports global CSS — [locale]/layout.tsx is now the effective root.

Result: 0 -> 5274 prerendered routes, 0 -> 70 ISR fallback patterns.
Cache-Control flips to s-maxage=31536000 with x-nextjs-cache: HIT, so
Cloudflare can serve from edge cache without invoking the worker for
landing pages. Rendered HTML is byte-identical to before across 10
sampled pages (/, /about, /use-cases, /blog, /privacy, /learn,
/de/learn, /pricing, /contact, /docs).
